refactor: refactor `renameKeysWithPrefix` to use prefixKeysForDeletion
